+++ lib/qc/modules/qc_actions/l10n.dart
@@ -39,4 +39,8 @@ class L10n extends BaseL10n {
 
   String get scanDeviceBarcode => Intl.message('Scan Device Barcode', locale: localName, name: 'scanDeviceBarcode');
 
+  String get warehouseAudit => Intl.message('Warehouse Audit', locale: localName, name: 'warehouseAudit');
+
+  String get deviceDetails => Intl.message('Device Details', locale: localName, name: 'deviceDetails');
+
 }