+++ lib/qc/modules/warehouse_audit/widgets/on_going_audit_widget.dart
@@ -1,0 +1,90 @@
+import 'package:core_widgets/core_widgets.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_trc/qc/modules/warehouse_audit/resources/ongoing_audit_response.dart';
+import 'package:flutter_trc/qc/modules/warehouse_audit/resources/warehouse_audit_service.dart';
+import 'package:flutter_trc/qc/modules/warehouse_audit/screens/warehouse_audit_perform_screen.dart';
+import 'package:flutter_trc/src/common/widgets/shimmer_list_widget.dart';
+
+import '../l10n.dart';
+
+class OnGoingAuditWidget extends StatefulWidget {
+  const OnGoingAuditWidget({super.key});
+
+  @override
+  State<OnGoingAuditWidget> createState() => _OnGoingAuditWidgetState();
+}
+
+class _OnGoingAuditWidgetState extends State<OnGoingAuditWidget> {
+  @override
+  Widget build(BuildContext context) {
+    var l10n = L10n(context);
+    return StreamBuilder(
+      stream: WarehouseAuditService.getOngoingAuditList(),
+      builder: (context, asyncSnapshot) {
+        if (asyncSnapshot.connectionState == ConnectionState.waiting) {
+          return const ShimmerListWidget();
+        }
+
+        if (Validator.isListNullOrEmpty(asyncSnapshot.data?.onGoingAuditList)) {
+          return Center(child: CshTextNew.subTitle1(l10n.emptyAuditList));
+        }
+
+        if (asyncSnapshot.hasData && asyncSnapshot.data != null) {
+          var list = asyncSnapshot.data?.onGoingAuditList;
+          return CshList(
+              rowCount: list?.length ?? 0,
+              onRefresh: () {
+                setState(() {});
+              },
+              getRowWidget: (index) {
+                var item = list?[index];
+                return GestureDetector(
+                  onTap: () => WarehouseAuditPerformScreen.pushNamed(context, item!.auditId!),
+                  child: _Item(item),
+                );
+              });
+        }
+        return const SizedBox.shrink();
+      },
+    );
+  }
+}
+
+class _Item extends StatelessWidget {
+  final OnGoingAuditData? item;
+
+  const _Item(this.item, {super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    var l10n = L10n(context);
+    return CshCard(
+      child: Column(
+        children: [
+          _row(l10n.auditId, item?.auditId.toString() ?? ""),
+          _row(l10n.facilityName, item?.facilityName ?? ""),
+          _row(
+            l10n.startTime,
+            formatDate(timeStamp: item?.startDate?.toInt(), pattern: DateFormats.dd_MMM_yyyy_HH_mm_ss.value),
+          ),
+          _row(
+            l10n.endTime,
+            formatDate(timeStamp: item?.endDate?.toInt(), pattern: DateFormats.dd_MMM_yyyy_HH_mm_ss.value),
+          ),
+        ],
+      ),
+    );
+  }
+
+  _row(String label, String? value) {
+    return Padding(
+      padding: const EdgeInsets.symmetric(vertical: Dimens.space_4),
+      child: Row(
+        children: [
+          Flexible(flex: 2, fit: FlexFit.tight, child: CshTextNew.subTitle2(label, isPrimary: false)),
+          Flexible(flex: 3, fit: FlexFit.tight, child: CshTextNew.subTitle1(value ?? "")),
+        ],
+      ),
+    );
+  }
+}