+++ lib/qc/qc_routes.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_trc/qc/modules/device_details/screens/device_details_screen.dart';
 import 'package:flutter_trc/qc/modules/external_audit/external_audit_home_screen.dart';
 import 'package:flutter_trc/qc/modules/external_audit/external_audit_perform_screen.dart';
 import 'package:flutter_trc/qc/modules/gaurd/screens/guard_device_counting_list_screen.dart';
@@ -21,8 +22,11 @@ import 'package:flutter_trc/qc/modules/stock_transfer/screens/pending_dispatch_d
 import 'package:flutter_trc/qc/modules/stock_transfer/screens/pending_lot_detail_screen.dart';
 import 'package:flutter_trc/qc/modules/stock_transfer/screens/st_store_out_screen.dart';
 import 'package:flutter_trc/qc/modules/stock_transfer/screens/stock_transfer_list_screen.dart';
+import 'package:flutter_trc/qc/modules/stock_transfer/screens/storage_device_list_screen.dart';
 import 'package:flutter_trc/qc/modules/store_in/screens/index.dart';
 import 'package:flutter_trc/qc/modules/supervisor/screens/supervisor_screen.dart';
+import 'package:flutter_trc/qc/modules/warehouse_audit/screens/on_going_audit_screen.dart';
+import 'package:flutter_trc/qc/modules/warehouse_audit/screens/warehouse_audit_perform_screen.dart';
 import 'package:flutter_trc/qc/qc_common/lot_type_filters/screens/store_out_lot_filter_screen.dart';
 import 'package:flutter_trc/src/common/gallery_screen.dart';
 import 'package:flutter_trc/src/common/user/widget/user_profile_screen.dart';
@@ -90,6 +94,10 @@ class QcRoutes {
       UserProfileScreen.route: (_) => const UserProfileScreen(),
       SupervisorScreen.route: (_) => const SupervisorScreen(),
       GalleryScreen.route: (_) => const GalleryScreen(),
+      OnGoingAuditScreen.route: (_) => const OnGoingAuditScreen(),
+      WarehouseAuditPerformScreen.route: (_) => const WarehouseAuditPerformScreen(),
+      DeviceDetailsScreen.route: (_) => const DeviceDetailsScreen(),
+      StorageDeviceListScreen.route: (_) => const StorageDeviceListScreen(),
     };
   }
 }