+++ lib/src/libraries/analytics/events/calculator_answer_selected_event.dart
@@ -1,0 +1,31 @@
+import 'package:flutter_trc/src/common/calculator_analytics/calculator_question_answer_request.dart';
+import 'package:flutter_trc/src/libraries/analytics/analytic_event_keys.dart';
+import 'package:flutter_trc/src/libraries/analytics/analytic_event_params.dart';
+import 'package:flutter_trc/src/libraries/analytics/events/common_events.dart';
+
+class CalculatorAnswerSelectedEvent extends CommonEvents {
+  final String? barcode;
+  final List<CalculatorQuestionAnswerRequest>? questionAnswerRequest;
+  final int pageNo;
+
+  CalculatorAnswerSelectedEvent(this.barcode, this.pageNo, this.questionAnswerRequest);
+
+  @override
+  String getKey() {
+    String eventName = "${AnalyticEventKeys.manualTesting.calWithPageNo}$pageNo";
+    return eventName;
+  }
+
+  @override
+  String getSubordinateKey() {
+    return AnalyticEventKeys.manualTesting.subOrdinateKey;
+  }
+
+  @override
+  Future<Map<String, dynamic>?> getArguments() async {
+    var arg = await super.getArguments();
+    arg?[AnalyticEventParams.deviceBarcode] = barcode;
+    arg?[AnalyticEventParams.metaData] = questionAnswerRequest;
+    return arg;
+  }
+}