+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # TRC Releases
 
 
+# v3.1.0+31 (02-02-2024)
+
+### Features And Improvements:
+
+* `QC - Stock Transfer - Store-out Pending Transfer Lot logic changes`[(SCM-5158)](https://reglobe.atlassian.net/browse/SCM-5158)
+* `QC - Stock Transfer - Stop remove and add device after transfer status lock`[(SCM-5341)](https://reglobe.atlassian.net/browse/SCM-5341)
+* `QC - Device Status - create new page to show device details`[(SCM-5407)](https://reglobe.atlassian.net/browse/SCM-5407)
+* `QC - Audit Warehouse`[(SCM-5450)](https://reglobe.atlassian.net/browse/SCM-5450)
+* `QC - Manual Testing - Add Analytics`[(SCM-5458)](https://reglobe.atlassian.net/browse/SCM-5458)
+
+
 # v3.0.10+30 (23-01-2024)
 
 ### Bug Fix: